# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FloraSync API",
    "version": "1.0.0",
    "description": "Local backend API for FloraSync greenhouse management."
  },
  "paths": {
...
    "/api/gardens": {
      "get": {
        "summary": "Admin Endpoint to Get All Gardens",
        "tags": ["Gardens"],
        "responses": {
          "200": { "description": "List of all gardens retrieved successfully" },
          "403": { "description": "Forbidden - Only admins can view all gardens." },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/gardens/{id}": {
      "put": {
        "summary": "Admin Endpoint to Rename Garden",
        "tags": ["Gardens"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Garden renamed successfully" },
          "400": { "description": "Bad Request - Name required" },
          "403": { "description": "Forbidden - Only admins can rename gardens." },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/workspaces": {
      "get": {
        "summary": "Get all gardens the user has access to",
        "tags": ["Workspaces"],
        "responses": {
          "200": { "description": "List of accessible workspaces retrieved successfully" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/garden/profile": {
      "put": {
        "summary": "Update shared garden profile info",
        "tags": ["Gardens"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "imageUrl": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Garden profile updated successfully" },
          "400": { "description": "Bad Request - No garden assigned" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/state": {
      "get": {
        "summary": "Fetch entire garden state",
        "tags": ["State"],
        "parameters": [
          {
            "name": "gardenId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Complete garden state retrieved successfully" },
          "403": { "description": "Forbidden - Access denied to this garden" },
          "500": { "description": "Internal server error" }
        }
      },
      "post": {
        "summary": "Update garden state",
        "tags": ["State"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "instances": { "type": "array", "items": {} },
                  "locations": { "type": "array", "items": {} },
                  "zones": { "type": "array", "items": {} },
                  "printQueue": { "type": "array", "items": {} },
                  "gardenJournal": { "type": "array", "items": {} },
                  "archetypes": { "type": "array", "items": {} }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Garden state updated successfully" },
          "403": { "description": "Forbidden - Viewers cannot modify state" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/api/gardens/action/rain": {
      "post": {
        "summary": "Bulk Action Endpoint: Log Rain",
        "tags": ["Gardens"],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rainType": { "type": "string", "example": "Light Sprinkle" },
                  "duration": { "type": "string", "example": "1 hr 15 mins" },
                  "durationMinutes": { "type": "integer", "example": 75 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Rain logged successfully" },
          "403": { "description": "Forbidden - Viewers cannot modify state" },
          "404": { "description": "Not Found - Garden not found" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
  }
}